requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
pymupdf>=1.23.6
python-docx>=1.1.0
presidio-analyzer>=2.2.351
presidio-anonymizer>=2.2.351
spacy>=3.7.2
neo4j
llama-index
llama-index-graph-stores-neo4j
llama-index-llms-gemini
python-dotenv
pydantic-settings
google-generativeai
requests
